+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 **Dev**
+- Move function utilities to tools.py
 
 **0.2.0**
 - Fix error decorator for Sphinx doc generation